requirements.txt
Django==4.2
firebase-admin==6.2.0
google-generativeai==0.3.1
python-dotenv==1.0.1
requests==2.31.0
